Refs 12540. Moved IReaderDataFilter to base rtps folder.
